requirements.txt
ollama==0.1.5
PyPDF2==3.0.1
python-docx==0.8.11